requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-ckeditor==5.9.0
django-heroku==0.3.1
django-import-export==2.0.2
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
mysql==0.0.2
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
postgres==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
static3==0.7.0
tablib==1.1.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
